% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_json_with_schema}
\alias{validate_json_with_schema}
\title{Validate a JSON string against a JSON Schema string}
\usage{
validate_json_with_schema(json_string, schema_string)
}
\arguments{
\item{json_string}{The JSON code}

\item{schema_string}{The JSON Schema string
This function will check that the supplied schema is not empty and it will then validate 
that the JSON code is valid (and well formed) w.r.t the supplied schema.}
}
\description{
Description
}

