% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_json_with_schemafile}
\alias{validate_json_with_schemafile}
\title{Validate a JSON string against a JSON Schema file}
\usage{
validate_json_with_schemafile(json_string, schemafn)
}
\arguments{
\item{json_string}{The JSON code}

\item{schemafn}{The JSON Schema file
This function will check that the supplied schema file is not empty and is accessible
and it will then validate that the JSON code is valid (and well formed) against the
supplied schema file.}
}
\description{
Description
}
\examples{
\dontrun{
json_code <- "{\\"category\\": \\"book\\", \\"price\\": 25,  \\"title\\": \\"abrakadabra\\"}"
validate_json_with_schemafile(json_code, "data/schema.json")
}
}

