% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_jsonfile_with_schema}
\alias{validate_jsonfile_with_schema}
\title{Validate a JSON file against a JSON Schema string}
\usage{
validate_jsonfile_with_schema(jsonfn, schema_string)
}
\arguments{
\item{jsonfn}{The JSON file}

\item{schema_string}{The JSON Schema string
This function will check that the supplied schema is not empty and it will then validate
that the JSON file is accessible and valid (and well formed) w.r.t the supplied schema file.}
}
\description{
Description
}
\examples{
\dontrun{
json_code <- "{\\"category\\": \\"book\\", \\"price\\": 25,  \\"title\\": \\"abrakadabra\\"}"
validate_json_with_schemafile(json_code, "data/schema.json")
}
}

