% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_jsonfile_with_schemafile}
\alias{validate_jsonfile_with_schemafile}
\title{Validate a JSON file against a JSON Schema file}
\usage{
validate_jsonfile_with_schemafile(jsonfn, schemafn)
}
\arguments{
\item{jsonfn}{The JSON file}

\item{schemafn}{The JSON Schema file
This function will check that the supplied files are not empty and are accessible
and it will then validate that the JSON file is valid (and well formed) against the
supplied schema file. Currently only version 4 of the JSON schema is supported.}
}
\description{
Description
}
\examples{
\dontrun{
validate_jsonfile_with_schemafile("data/item-3.json", "data/schema.json")
}
}

